Extracted side bar to own component
