+++ src/client/layouts/user/UserLayout.tsx
@@ -1,14 +1,11 @@
-import React, { useEffect, useState, useContext } from 'react';
+import React, { useEffect, useContext } from 'react';
 import { Outlet, useNavigate } from 'react-router-dom';
 
 // Types
-import {
-  UserSideBarProps,
-  UserContextValue,
-  PuzzleCollectionContextValue
-} from '../../frontendTypes';
+import { UserContextValue, PuzzleCollectionContextValue } from '../../frontendTypes';
 
 // Components
+import SideBarContainer from '../../shared-components/SideBarContainer';
 import UserSideBar from './components/UserSideBar';
 
 // Context
@@ -20,7 +17,6 @@ import signInWithSession from '../../utils/signInWithSession';
 // Main Component
 const UserLayout = () => {
   const navigate = useNavigate();
-  const [isSideBarExpanded, setIsSideBarExpanded] = useState(false);
   const { user, setUser } = useContext<UserContextValue>(userContext);
   const { setPuzzleCollection } = useContext<PuzzleCollectionContextValue>(puzzleCollectionContext);
 
@@ -34,45 +30,13 @@ const UserLayout = () => {
     }
   }, [user, setUser, setPuzzleCollection, navigate]);
 
-  const switchSideBarExpanded = () => {
-    setIsSideBarExpanded(!isSideBarExpanded);
-  };
-
-  // If user clicks on something other than a navbar child link, collapse the nav bar
-  // Each navlink also collapses the navbar, but the click has to register before the navbar
-  // collapses. It's done this way instead of just collapsing on any blur as the onBlur event
-  // is triggered before the navlink click, which prevented the navlink click from processing
-  const exteriorBlurCollapseSideBar = (e: React.FocusEvent) => {
-    if (!e.currentTarget.contains(e.relatedTarget)) {
-      setIsSideBarExpanded(false);
-    }
-  };
-
-  const userSideBarProps: UserSideBarProps = {
-    collapseSideBar() {
-      setIsSideBarExpanded(false);
-    }
-  };
-
   return (
-    <div id='user-layout'>
-      <div
-        className='user-side-bar-container'
-        tabIndex={0}
-        role='toolbar'
-        onBlur={(e) => exteriorBlurCollapseSideBar(e)}
-      >
-        <button id='side-bar-button' onClick={switchSideBarExpanded}>
-          <img src='/assets/list.svg' alt='list icon'></img>
-        </button>
-        <div className={`user-side-bar ${isSideBarExpanded ? '' : 'inactive'}`}>
-          <UserSideBar key='UserSideBar' {...userSideBarProps} />
-        </div>
-      </div>
+    <>
+      <SideBarContainer SideBar={UserSideBar} />
       <main>
         <Outlet />
       </main>
-    </div>
+    </>
   );
 };
